Table component/types cleanup
